requirements.txt
asgiref==3.2.10
Django==3.0.8
django-getenv==1.3.2
pytz==2020.1
sqlparse==0.3.1
